refactor: split approval UI and runtime helpers
