+++ container/src/index.ts
@@ -1,10 +1,4 @@
 import path from 'node:path';
-import { resolveBorderlineAnomalyWithTraceJudge } from './anomaly-trace-judge.js';
-import {
-  type ApprovalPrelude,
-  type ToolApprovalEvaluation,
-  TrustedAgentApprovalRuntime,
-} from './approval-policy.js';
 import { discoverArtifactsSince, inferArtifactMimeType } from './artifacts.js';
 import {
   cleanupAllBrowserSessions,
@@ -73,6 +67,16 @@ import {
   readChatCompletionUsageTokens,
   recordPerformanceSample,
 } from './token-usage.js';
+import {
+  type ApprovalPrelude,
+  approvalRuntime,
+  buildApprovalDeniedToolExecution,
+  buildApprovalRequiredToolExecution,
+  buildPendingApproval,
+  createToolApprovalResolver,
+  emitApprovalProgress,
+  type ToolApprovalEvaluation,
+} from './tool-approval.js';
 import { parseToolArgsJson } from './tool-args.js';
 import { validateStructuredToolCalls } from './tool-call-validation.js';
 import type { ToolCallHistoryEntry } from './tool-loop-detection.js';
@@ -114,7 +118,6 @@ import {
   type ContainerInput,
   type ContainerOutput,
   type EscalationTarget,
-  type PendingApproval,
   TASK_MODEL_KEYS,
   type ToolCall,
   type ToolDefinition,
@@ -159,17 +162,6 @@ function applyRuntimeEnv(runtimeEnv: ContainerInput['runtimeEnv']): void {
   }
 }
 
-const approvalRuntime = new TrustedAgentApprovalRuntime();
-approvalRuntime.setApprovalRuleHookEmitter((event) =>
-  emitRuntimeEvent({
-    event: event.hook,
-    kind: event.kind,
-    approvalRule: event.ruleName,
-    toolName: event.toolName,
-    ...(event.actionKey ? { actionKey: event.actionKey } : {}),
-    ...(event.decision ? { decision: event.decision } : {}),
-  }),
-);
 let cachedSelectedSkillPath: string | null = null;
 
 /** Auth material received once via stdin, held in memory for the agent lifetime. */
@@ -396,43 +388,6 @@ function emitStreamThinkingDelta(delta: string): void {
 
 function emitStreamActivity(): void {
   console.error('[stream-activity]');
-}
-
-function emitApprovalProgress(approval: PendingApproval): void {
-  const payload = Buffer.from(JSON.stringify(approval), 'utf-8').toString(
-    'base64',
-  );
-  console.error(`[approval] ${payload}`);
-}
-
-function buildPendingApproval(
-  approval: ToolApprovalEvaluation,
-  prompt: string,
-  toolName: string,
-): PendingApproval {
-  if (!approval.requestId) {
-    throw new Error('Approval-required tool call is missing a request id.');
-  }
-  return {
-    approvalId: approval.requestId,
-    prompt,
-    intent: approval.intent,
-    reason: approval.reason,
-    approvalTier: approval.tier,
-    toolName,
-    commandPreview: approval.commandPreview,
-    allowSession: !approval.pinned,
-    allowAgent: !approval.pinned,
-    allowAll: !approval.pinned,
-    expiresAt:
-      typeof approval.expiresAtMs === 'number' &&
-      Number.isFinite(approval.expiresAtMs)
-        ? approval.expiresAtMs
-        : null,
-    ...(approval.escalationTarget
-      ? { escalationTarget: approval.escalationTarget }
-      : {}),
-  };
 }
 
 function latestUserPrompt(messages: ChatMessage[]): string {
@@ -1134,60 +1089,29 @@ async function processRequest(
     return output;
   }
 
-  const resolveToolApproval = async (input: {
-    toolName: string;
-    argsJson: string;
-  }): Promise<ToolApprovalEvaluation> => {
-    const approvalEvaluatedAt = new Date();
-    let evaluation = approvalRuntime.evaluateToolCall({
-      toolName: input.toolName,
-      argsJson: input.argsJson,
-      latestUserPrompt: effectiveUserPrompt,
-      channelId,
-      escalationTarget,
-      now: approvalEvaluatedAt,
-    });
-    const resolved = await resolveBorderlineAnomalyWithTraceJudge({
-      evaluation,
-      toolName: input.toolName,
-      argsJson: input.argsJson,
-      latestUserPrompt: effectiveUserPrompt,
-      taskModels,
-      fallbackContext: {
-        provider,
-        providerMethod,
-        baseUrl,
-        apiKey,
-        model,
-        chatbotId,
-        requestHeaders,
-        isLocal,
-        contextWindow,
-        thinkingFormat,
-        debugModelResponses,
-      },
-    });
-    if (resolved.response) {
+  const resolveToolApproval = createToolApprovalResolver({
+    latestUserPrompt: effectiveUserPrompt,
+    channelId,
+    escalationTarget,
+    taskModels,
+    fallbackContext: {
+      provider,
+      providerMethod,
+      baseUrl,
+      apiKey,
+      model,
+      chatbotId,
+      requestHeaders,
+      isLocal,
+      contextWindow,
+      thinkingFormat,
+      debugModelResponses,
+    },
+    onModelResponse: (response) => {
       tokenUsage.modelCalls += 1;
-      accumulateApiUsage(tokenUsage, resolved.response);
-    }
-    const traceJudge = resolved.evaluation.anomaly?.traceJudge;
-    if (evaluation.anomaly?.tuple && traceJudge) {
-      approvalRuntime.recordAnomalyTraceJudgeResult(
-        evaluation.anomaly.tuple,
-        traceJudge,
-      );
-      evaluation = approvalRuntime.evaluateToolCall({
-        toolName: input.toolName,
-        argsJson: input.argsJson,
-        latestUserPrompt: effectiveUserPrompt,
-        channelId,
-        escalationTarget,
-        now: approvalEvaluatedAt,
-      });
-    }
-    return evaluation;
-  };
+      accumulateApiUsage(tokenUsage, response);
+    },
+  });
 
   if (approvedToolCall) {
     const approval = await resolveToolApproval({
@@ -1206,29 +1130,12 @@ async function processRequest(
         result: prompt,
         toolsUsed: [approvedToolCall.toolName],
         toolExecutions: [
-          {
-            name: approvedToolCall.toolName,
-            arguments: approvedToolCall.argsJson,
-            result: prompt,
-            durationMs: 0,
-            isError: false,
-            blocked: true,
-            blockedReason: approval.reason,
-            approvalTier: approval.tier,
-            approvalBaseTier: approval.baseTier,
-            autonomyLevel: approval.autonomyLevel,
-            stakes: approval.stakes,
-            stakesScore: approval.stakesScore,
-            anomaly: approval.anomaly,
-            escalationRoute: approval.escalationRoute,
-            escalationTarget: approval.escalationTarget,
-            approvalDecision: approval.decision,
-            approvalActionKey: approval.actionKey,
-            approvalIntent: approval.intent,
-            approvalReason: approval.reason,
-            approvalRequestId: approval.requestId,
-            approvalExpiresAt: approval.expiresAtMs,
-          },
+          buildApprovalRequiredToolExecution({
+            toolName: approvedToolCall.toolName,
+            argsJson: approvedToolCall.argsJson,
+            prompt,
+            approval,
+          }),
         ],
         pendingApproval,
         tokenUsage: finalizeTokenUsage(tokenUsage),
@@ -1777,32 +1684,14 @@ async function processRequest(
           toolName,
         );
         emitApprovalProgress(pendingApproval);
-        toolExecutions.push({
-          name: toolName,
-          arguments: call.function.arguments,
-          result: prompt,
-          durationMs: 0,
-          isError: false,
-          blocked: true,
-          blockedReason: approval.reason,
-          approvalTier: approval.tier,
-          approvalBaseTier: approval.baseTier,
-          autonomyLevel: approval.autonomyLevel,
-          stakes: approval.stakes,
-          stakesScore: approval.stakesScore,
-          anomaly: approval.anomaly,
-          escalationRoute: approval.escalationRoute,
-          escalationTarget: approval.escalationTarget,
-          approvalDecision: approval.decision,
-          approvalActionKey: approval.actionKey,
-          approvalIntent: approval.intent,
-          approvalReason: approval.reason,
-          approvalRequestId: approval.requestId,
-          approvalExpiresAt: approval.expiresAtMs,
-          approvalAllowSession: !approval.pinned,
-          approvalAllowAgent: !approval.pinned,
-          approvalAllowAll: !approval.pinned,
-        });
+        toolExecutions.push(
+          buildApprovalRequiredToolExecution({
+            toolName,
+            argsJson: call.function.arguments,
+            prompt,
+            approval,
+          }),
+        );
         const waitingForApproval: ContainerOutput = {
           status: 'success',
           result: prompt,
@@ -1824,31 +1713,14 @@ async function processRequest(
       if (approval.decision === 'denied') {
         toolsUsed.push(toolName);
         const denialText = `Approval denied: ${approval.reason}`;
-        toolExecutions.push({
-          name: toolName,
-          arguments: call.function.arguments,
-          result: denialText,
-          durationMs: 0,
-          isError: true,
-          blocked: true,
-          blockedReason: approval.reason,
-          approvalTier: approval.tier,
-          approvalBaseTier: approval.baseTier,
-          autonomyLevel: approval.autonomyLevel,
-          stakes: approval.stakes,
-          stakesScore: approval.stakesScore,
-          anomaly: approval.anomaly,
-          escalationRoute: approval.escalationRoute,
-          escalationTarget: approval.escalationTarget,
-          approvalDecision: approval.decision,
-          approvalActionKey: approval.actionKey,
-          approvalIntent: approval.intent,
-          approvalReason: approval.reason,
-          approvalRequestId: approval.requestId,
-          approvalExpiresAt: approval.expiresAtMs,
-          approvalAllowSession: !approval.pinned,
-          approvalAllowAgent: !approval.pinned,
-        });
+        toolExecutions.push(
+          buildApprovalDeniedToolExecution({
+            toolName,
+            argsJson: call.function.arguments,
+            denialText,
+            approval,
+          }),
+        );
         const denied: ContainerOutput = {
           status: 'success',
           result: denialText,
